worker_photo/requirements.txt
pika==1.2.0
pika-wrapper==0.0.3
requests==2.26.0
Pillow==8.3.1
pyTelegramBotAPI==3.8.1
